apps/api/requirements.txt
fastapi
uvicorn[standard]
pymongo
langchain
python-dotenv
openai